# Varargs don't specialize