oops.
Before:
```
julia> @benchmark sin_sum(x, y, z) setup=(x=rand(); y = rand(); z = rand())
BenchmarkTools.Trial: 10000 samples with 990 evaluations.
 Range (min … max):  44.685 ns …  1.015 μs  ┊ GC (min … max): 0.00% … 92.17%
 Time  (median):     48.480 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   52.340 ns ± 35.280 ns  ┊ GC (mean ± σ):  2.69% ±  4.03%

   ▁█▅▆▅▂ ▁                                                    
  ▂████████▇▅▆▄▄▃▃▃▃▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  44.7 ns         Histogram: frequency by time        80.2 ns <

 Memory estimate: 96 bytes, allocs estimate: 5.
```
After:
```
julia> @benchmark sin_sum(x, y, z) setup=(x=rand(); y = rand(); z = rand())
BenchmarkTools.Trial: 10000 samples with 999 evaluations.
 Range (min … max):   9.629 ns … 144.460 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     11.763 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   11.850 ns ±   2.224 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

               █ ▇      ▃ ▅ ▁                                   
  ▂▁▂▂▃▂▄▃▃▄▃█▂█▂█▂█▄▃▆▂█▃█▃█▅▄▅▂▄▂▄▂▄▃▃▅▂▄▂▄▂▃▂▃▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  9.63 ns         Histogram: frequency by time         15.4 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```